+++ open-ai-utils/src/request_builder/other_request_data.rs
@@ -1,9 +1,9 @@
 #[derive(Debug, Clone, Default)]
 pub struct OtherRequestData {
-    pub n: Option<u32>,
+    pub n: Option<i64>,
     pub presence_penalty: Option<f64>,
     pub frequency_penalty: Option<f64>,
     pub top_p: Option<f64>,
-    pub top_k: Option<u32>,
+    pub top_k: Option<i64>,
     pub temperature: Option<f64>,
 }
